requirements.txt
streamlit>=1.28.0
gradio>=4.0.0
ollama-python>=0.1.0
faiss-cpu>=1.7.4
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
sqlalchemy>=2.0.0
